@@ -1,6 +1,6 @@
 {
   "name": "@react-hive/honey-style",
-  "version": "1.3.0",
+  "version": "1.4.0",
   "description": "",
   "keywords": [
     "react",
@@ -19,12 +19,14 @@
   "license": "MIT",
   "scripts": {
     "prepare": "husky",
-    "clean": "rm -rf dist",
-    "build": "NODE_ENV=production webpack --config webpack.config.js",
+    "build": "NODE_ENV=production webpack --config webpack.config.mjs",
     "test": "jest --collect-coverage",
     "diagnostics": "tsx scripts/collect-diagnostics.ts"
   },
-  "main": "dist/index.js",
+  "type": "module",
+  "exports": {
+    ".": "./dist/index.mjs"
+  },
   "types": "dist/index.d.ts",
   "files": [
     "dist",
@@ -41,16 +43,16 @@
     "stylis": "4.3.6"
   },
   "devDependencies": {
-    "@eslint/js": "9.26.0",
+    "@eslint/js": "9.27.0",
     "@mdx-js/loader": "3.1.0",
     "@testing-library/jest-dom": "6.6.3",
     "@testing-library/react": "16.3.0",
     "@types/jest": "29.5.14",
     "@types/mdx": "2.0.13",
     "@types/node": "22.15.17",
-    "@types/react": "^19.1.3",
+    "@types/react": "^19.1.4",
     "@types/stylis": "4.2.7",
-    "eslint": "9.26.0",
+    "eslint": "9.27.0",
     "eslint-plugin-react": "7.37.5",
     "html-webpack-plugin": "5.6.3",
     "husky": "9.1.7",
@@ -61,7 +63,7 @@
     "ts-loader": "9.5.2",
     "tsx": "4.19.4",
     "typescript": "5.8.3",
-    "typescript-eslint": "8.32.0",
+    "typescript-eslint": "8.32.1",
     "webpack": "5.99.8",
     "webpack-cli": "6.0.1"
   },
