@@ -1,6 +1,10 @@
+import path, { dirname } from 'path';
+import { fileURLToPath } from 'url';
 import { execSync } from 'child_process';
 import fs from 'fs';
-import path from 'path';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 const toCamelCase = (str: string): string =>
   str
