@@ -2,13 +2,15 @@ import type { SVGAttributes, TdHTMLAttributes } from 'react';
 
 import type { HoneyBreakpointName } from './types';
 
-export const ENV =
-  typeof process !== 'undefined' && process.env?.NODE_ENV ? process.env.NODE_ENV : 'development';
+const ENV =
+  (typeof process !== 'undefined' && typeof process.env !== 'undefined' && process.env.NODE_ENV) ||
+  'development';
 
 export const __DEV__ = ENV !== 'production';
 
 export const HONEY_STYLED_COMPONENT_ID_PROP = '$$ComponentId';
 
+export const HONEY_GLOBAL_STYLE_ATTR = 'data-honey-global-style';
 export const HONEY_STYLE_ATTR = 'data-honey-style';
 
 export const HONEY_BREAKPOINTS: HoneyBreakpointName[] = ['xs', 'sm', 'md', 'lg', 'xl'];
